// crates/lite_mobile/ui/src/lib/models/rustModels/ReportStep.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ReportInfo } from "./ReportInfo";

export type ReportStep = { "FnsReports": { text: ReportInfo, } } | { "HomeReports": { text: ReportInfo, } } | { "Loading": { text: ReportInfo, } } | { "SaveFiles": { text: ReportInfo, xls_file: Array<number>, pdf_file: Array<number>, } } | { "TryLater": { text: ReportInfo, } };